+++ test_helpers/main.go
@@ -83,7 +83,6 @@ type TarantoolInstance struct {
 func isReady(dialer tarantool.Dialer, opts *tarantool.Opts) error {
 	var err error
 	var conn *tarantool.Connection
-	var resp tarantool.Response
 
 	ctx, cancel := GetConnectContext()
 	defer cancel()
@@ -96,12 +95,9 @@ func isReady(dialer tarantool.Dialer, opts *tarantool.Opts) error {
 	}
 	defer conn.Close()
 
-	resp, err = conn.Do(tarantool.NewPingRequest()).GetResponse()
+	_, err = conn.Do(tarantool.NewPingRequest()).Get()
 	if err != nil {
 		return err
-	}
-	if resp == nil {
-		return errors.New("response is nil after ping")
 	}
 
 	return nil