+++ test_helpers/request_mock.go
@@ -2,6 +2,7 @@ package test_helpers
 
 import (
 	"context"
+	"io"
 
 	"github.com/tarantool/go-iproto"
 	"github.com/vmihailenco/msgpack/v5"
@@ -34,4 +35,11 @@ func (sr *StrangerRequest) Conn() *tarantool.Connection {
 
 func (sr *StrangerRequest) Ctx() context.Context {
 	return nil
+}
+
+func (sr *StrangerRequest) Response(header tarantool.Header,
+	body io.Reader) (tarantool.Response, error) {
+	resp := tarantool.BaseResponse{}
+	resp.SetHeader(header)
+	return &resp, nil
 }